refactor(ui-watt): create picker base-class to reduce code duplications
